fix: typing & add open/close state
